Add support for an "_" key which appends values to the end of the argument array
Add support for an "_" key which appends values to the end of the argument array. This
imitates [`minimist`](https://github.com/substack/minimist#example), which includes
end arguments with no options as an array of values keyed by "_".

If "_"'s value is not an Array, a TypeError will be thrown.

Updated tests and readme to reflect changes.