Issue #16 optional mapping of keys to short arguments
A generic solution to close #16. Allow the user to pass in
an `aliases` object in the `options`. Each entry maps an
accepted value to a short argument. For example, passing

`dargs({file: 'value'}, {aliases: {file: 'f'}})`

might produce an output like

`[ '-f value']`

This is nice because it allows you to work with some commands
that only support short arg versions of certain arguments (I
was working with `ctags -f`), while still encouraging you to
expose expressive, full word arguments to the end user.

Updated tests with usage alone and with `includes`, and docs
with examples and short description.