includes and exludes accept regular expressions in addition to strings
The `includes` and `excludes` options now accept and handle regular
expressions in additions to strings. This makes including and excluding
keys much more flexible.

For example, a user providing 'pythonKinds' would match the includes
array `[/.*Kinds/]`. This allows for cases where the first part of a
command line argument is actually a hidden parameter. See `<LANG>-kinds`
in ctags for an example.

Augmented `includes` and `excludes` tests to include regular expressions,
and updated readme to include regex examples.