Treat elements of includes and excludes as globs instead of strings
Strings without special characters like parentheses, +-, and *, will
be treated the same.  Strings with glob characters will now match
any key provided by the user that matches the glob. The functionality
is provided by the minimatch package, which is no included in the
package.json.

For example, a user providing 'pythonKinds' would match the includes
array `['*Kinds']`. This allows for cases where the first part of a
command line argument is actually a hidden parameter. See `<LANG>-kinds`
in ctags for an example.

Augmented `includes` and `excludes` tests to include globs, and updated
readme to include glob examples.