Allow the user to configure the separator between command line arguments
Tries to fix #25 in a general way. Instead of the user being able to toggle the
'=' sign on and off, they can set the separator to any value. For example,
`dargs({foo: bar}, {separator: ' '})` produces `['foo bar']`.

Not sure if this is overkill, or less intuitive than the binary option.

Added tests and updated readme to reflect changes.